+++ src/app/pages/about.jsx
@@ -1,10 +1,7 @@
 import React from 'react';
 
-import { Nav } from '../components/header';
-
 const About = () => (
   <div>
-    <Nav />
     <div>About Me</div>
   </div>
 );