+++ src/app/pages/home.jsx
@@ -1,11 +1,9 @@
 import React from 'react';
-import { Nav } from '../components/header';
 import { PostList } from '../components/post';
 
 
 export default () => (
   <div>
-    <Nav />
     <PostList />
   </div>
 );