+++ src/app/routes.jsx
@@ -3,6 +3,7 @@ import createHistory from 'history/createBrowserHistory';
 import { Router, Route } from 'react-router-dom';
 import { routerMiddleware } from 'react-router-redux';
 
+import { Nav } from './components/header';
 import Home from './pages/home';
 import About from './pages/about';
 
@@ -13,6 +14,7 @@ const middleware = routerMiddleware(history);
 const Routes = () => (
   <Router history={history}>
     <div>
+      <Nav />
       <Route exact path="/" component={Home} />
       <Route path="/about" component={About} />
     </div>